requirements.txt
textual
playwright
beautifulsoup4
requests
openpyxl
PyYAML
